backend/requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiosmtpd==1.4.4.post2
alabaster==0.7.13
anyio==3.6.2
APScheduler==3.10.1
arrow==0.17.0
asgiref==3.6.0
astroid==2.4.2
async-timeout==4.0.2
atpublic==3.1.1
attrs==23.1.0
autopep8==1.5.4
Babel==2.12.1
beautifulsoup4==4.9.3
binaryornot==0.4.4
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cookiecutter==1.7.2
cookiejar==0.0.3
cryptography==40.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==4.0.0
Django==4.1.7
django-allauth==0.54.0
django-apscheduler==0.6.2
django-cors-headers==4.3.1
django-cron==0.6.0
django-crontab==0.7.1
django-formtools==2.4.1
django-jazzmin==2.6.0
django-otp==1.2.0
django-phone-verify==3.0.0
django-phonenumber-field==6.4.0
django-rest-framework==0.1.0
django-socialregistration==0.5.10
django-two-factor-auth==1.15.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docutils==0.20.1
fastapi==0.92.0
freetype-py==2.4.0
frozenlist==1.3.3
h11==0.14.0
html5lib==1.1
httpie==3.2.1
httplib2==0.22.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.6.0
isort==5.7.0
Jinja2==2.11.2
jinja2-time==0.2.0
joblib==1.3.1
lazy-object-proxy==1.4.3
markdown-it-py==2.2.0
MarkupSafe==1.1.1
mccabe==0.6.1
mdurl==0.1.2
mechanize==0.4.5
mock==5.0.2
MouseInfo==0.1.3
multidict==6.0.4
nexmo==2.5.2
nltk==3.8.1
numpy==1.24.3
oauth2==1.9.0.post1
oauthlib==3.2.2
packaging==23.1
pager==3.3
phonenumbers==8.13.11
Pillow==9.5.0
poyo==0.5.0
PyAutoGUI==0.9.52
pycodestyle==2.6.0
pycparser==2.21
pydantic==1.10.5
pygame==2.4.0
PyGetWindow==0.0.9
Pygments==2.15.1
PyJWT==2.7.0
pylibjpeg==1.4.0
pylint==2.6.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.1
pypng==0.20220715.0
PyRect==0.1.4
PyScreeze==0.1.26
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==1.0.0
python-openid==2.2.5
python-slugify==4.0.1
python-social-auth==0.3.6
python3-openid==3.2.0
PyTweening==1.0.3
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
regex==2023.6.3
requests==2.30.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.3.5
schedule==1.2.0
selenium==3.141.0
six==1.15.0
sniffio==1.3.0
snowballstemmer==2.2.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.1
Sphinx==7.0.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.3
starlette==0.25.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.65.0
twilio==8.2.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==5.0.1
urllib3==2.0.2
uvicorn==0.20.0
webencodings==0.5.1
wrapt==1.12.1
xmltodict==0.13.0
yarl==1.9.2
zipp==3.15.0
